backend/requirements.txt

fastapi==0.110.0
uvicorn==0.29.0
transformers==4.40.0
torch==2.2.1
tweepy==4.14.0
facebook-sdk==3.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.28
asyncpg==0.29.0
websockets==12.0
pydantic==2.6.2
